\name{WRScpp-package}
\Rdversion{1.1}
\alias{WRScpp-package}
\alias{WRScpp}
\docType{package}
\title{
  WRScpp subroutine
}
\description{
  C++ sub-routines for the R package WRS.
}
\author{
Robust functions by R.R. Wilcox
Package maintainer: Felix Schönbrodt <felix@nicebread.de>
}
\references{
Wilcox, R. R. (2005). \emph{Introduction to robust estimation and hypothesis testing.} San Diego: Academic Press.

}

\keyword{ package }
